+++ src/components/ingridientDitails/ingridientDitalis.jsx
@@ -22,3 +22,19 @@ function Ingridient(props) {
 }
 
 export default Ingridient;
+
+// import { CurrencyIcon } from "@ya.praktikum/react-developer-burger-ui-components";
+// function IngredientDetales(props) {
+//   return (
+//     <li>
+//       <img src={props.item.image} alt={props.item.name} />
+//       <div>
+//         <p>{props.item.price}</p>
+//         <CurrencyIcon type="primary" />
+//       </div>
+//       <p>{props.item.name}</p>
+//     </li>
+//   );
+// }
+
+// export default IngredientDetales;
